Downgraded DB schema to not include reference from User to created games
Reason: Everything was breaking because of it. Sufficient amount of DB
(re)engineering is required to make proper links possible. I have ideas
how to do it, but I do not want to spend time on it now.
I also commented out some tests that were failing because of this
relationship absence, and I modified GameService (+ it's tests)
to not try to use this relationship and only use correct existing one.